requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Markdown==3.7
multidict==6.2.0
orjson==3.10.16
packaging==24.0
pathlib2==2.3.7.post1
pillow==11.1.0
propcache==0.3.0
psutil==6.1.1
pycparser==2.22
pyee==11.1.1
Pygments==2.19.1
pyppeteer==2.0.0
pyproject_hooks==1.2.0
PyYAML==6.0.2
qqmusic-api-python==0.3.4
qrcode==8.0
qrcode-terminal==0.8
requests==2.32.3
schedule==1.2.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==1.26.20
websockets==10.4
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0
